requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Agent Orchestration
langchain>=0.3.8
langgraph>=0.2.60
langchain-openai>=0.2.8
langchain-core>=0.3.72

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0

# Session & Memory
redis==5.1.1

# Security
cryptography==43.0.3
PyJWT==2.9.0
passlib[bcrypt]==1.7.4

# Utilities
httpx==0.27.2
psycopg2-binary==2.9.11
